MOB-317 - Fixed PR comments
- Fixed problem with time zones to avoid being able to select wrong dates.

- Added EsmorgaTimePickerDialog to DS.

- Renamed isTimeSelected to onTimeSelected to avoid confusion.

- Created date formatting to put formatting functions in place for reuse throughout the app.
