+++ designsystem/src/main/java/cmm/apps/designsystem/EsmorgaDatePicker.kt
@@ -3,16 +3,16 @@ package cmm.apps.designsystem
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.DatePicker
 import androidx.compose.material3.DatePickerDefaults
-import androidx.compose.material3.DatePickerFormatter
 import androidx.compose.material3.DatePickerState
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.SelectableDates
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
-import java.util.concurrent.TimeUnit
+import java.time.Instant
+import java.time.LocalDate
+import java.time.ZoneId
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
@@ -29,10 +29,14 @@ fun EsmorgaDatePicker(state: DatePickerState) {
 }
 
 @OptIn(ExperimentalMaterial3Api::class)
-object PossibleSelectableDates: SelectableDates {
+object PossibleSelectableDates : SelectableDates {
     override fun isSelectableDate(utcTimeMillis: Long): Boolean {
-        val utcDay = TimeUnit.MILLISECONDS.toDays(utcTimeMillis)
-        val actualDay = TimeUnit.MILLISECONDS.toDays(System.currentTimeMillis())
-        return utcDay >= actualDay
+        val dateFromPicker = Instant.ofEpochMilli(utcTimeMillis)
+            .atZone(ZoneId.systemDefault())
+            .toLocalDate()
+
+        val today = LocalDate.now(ZoneId.systemDefault())
+
+        return !dateFromPicker.isBefore(today)
     }
 }
