+++ view/src/main/java/cmm/apps/esmorga/view/createeventdate/CreateEventFormDateViewModel.kt
@@ -4,6 +4,8 @@ import androidx.lifecycle.ViewModel
 import cmm.apps.esmorga.domain.event.model.CreateEventForm
 import cmm.apps.esmorga.view.createeventdate.model.CreateEventFormDateEffect
 import cmm.apps.esmorga.view.createeventdate.model.CreateEventFormDateUiState
+import cmm.apps.esmorga.view.dateformatting.DateFormatter.formatIsoDateTime
+import cmm.apps.esmorga.view.dateformatting.DateFormatter.formatTimeWithMillisUtcSuffix
 import kotlinx.coroutines.channels.BufferOverflow
 import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.flow.MutableStateFlow
@@ -11,11 +13,6 @@ import kotlinx.coroutines.flow.SharedFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asSharedFlow
 import kotlinx.coroutines.flow.asStateFlow
-import java.time.Instant
-import java.time.LocalDateTime
-import java.time.LocalTime
-import java.time.ZoneId
-import java.time.format.DateTimeFormatter
 import java.util.Date
 
 class CreateEventFormDateViewModel(
@@ -32,7 +29,7 @@ class CreateEventFormDateViewModel(
         _effect.tryEmit(CreateEventFormDateEffect.NavigateBack)
     }
 
-    fun isTimeSelected(selectedTime: String) {
+    fun onTimeSelected(selectedTime: String) {
         if (selectedTime.isNotEmpty()) {
             _uiState.value = _uiState.value.copy(
                 isButtonEnabled = true
@@ -41,21 +38,11 @@ class CreateEventFormDateViewModel(
     }
 
     fun formattedTime(hour: Int, minute: Int): String {
-        val formatter = DateTimeFormatter.ofPattern("HH:mm:ss.SSS'Z'")
-        val time = LocalTime.of(hour, minute).format(formatter)
-        return time
+        return formatTimeWithMillisUtcSuffix(hour, minute)
     }
 
     private fun formattedDateTime(date: Date, time: String) {
-        val instant: Instant = date.toInstant()
-        val systemZone: ZoneId = ZoneId.systemDefault()
-        val localDateTimeFromDate: LocalDateTime = instant.atZone(systemZone).toLocalDateTime()
-
-        val datePartFormatter = DateTimeFormatter.ISO_LOCAL_DATE
-        val datePart = localDateTimeFromDate.format(datePartFormatter)
-
-        val dateTime = "${datePart}T${time}"
-
+        val dateTime = formatIsoDateTime(date, time)
         _uiState.value = _uiState.value.copy(
             dateTime = dateTime
         )
