+++ view/src/test/java/cmm/apps/esmorga/view/screenshot/createEvent/CreateEventDateScreenScreenhotTest.kt
@@ -30,7 +30,7 @@ class CreateEventFormDateScreenScreenshotTest : BaseScreenshotTest() {
                     isButtonEnabled = isButtonEnabled,
                     onBackPressed = {},
                     onNextClick = { _, _ -> },
-                    isTimeSelected = { _ -> },
+                    onTimeSelected = { _ -> },
                     formattedTime = { _, _ -> "" }
                 )
             }
